// frontend-institution/src/@fake-db/types.d.ts
export type SearchItem = {
  id: number
  url: { name:string, params?: object}
  icon: string
  title: string
  category: string
}

export type SearchHeader = {
  header: string
  title:string
}